feat: move `socket` shim to socket module, add `TcpSocket::new`
