+++ src/shims/unix/mod.rs
@@ -23,6 +23,7 @@ pub use self::env::{EvalContextExt as _, UnixEnvVars};
 pub use self::fd::{EvalContextExt as _, UnixFileDescription};
 pub use self::fs::{DirTable, EvalContextExt as _};
 pub use self::mem::EvalContextExt as _;
+pub use self::socket::EvalContextExt as _;
 pub use self::socket_address::EvalContextExt as _;
 pub use self::sync::EvalContextExt as _;
 pub use self::tcp_socket::EvalContextExt as _;
